pyproject.toml
[tool.poetry]
name = "apitest"
version = "0.1.0"
description = ""
authors = ["Kauan Soares Mota <4kauanmota@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-pydantic-spec = "^0.5.0"
tinydb = "^4.8.0"


[tool.poetry.group.dev.dependencies]
httpie = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
